Extracted allocation methods to separate module
